Share more code. Charge batch balancer fee

+++ actors/miner/src/monies.rs
@@ -297,21 +297,21 @@ lazy_static! {
 }
 
 pub fn aggregate_prove_commit_network_fee(
-    aggregate_size: i64,
+    aggregate_size: usize,
     base_fee: &TokenAmount,
 ) -> TokenAmount {
     aggregate_network_fee(aggregate_size, &ESTIMATED_SINGLE_PROVE_COMMIT_GAS_USAGE, base_fee)
 }
 
 pub fn aggregate_pre_commit_network_fee(
-    aggregate_size: i64,
+    aggregate_size: usize,
     base_fee: &TokenAmount,
 ) -> TokenAmount {
     aggregate_network_fee(aggregate_size, &ESTIMATED_SINGLE_PRE_COMMIT_GAS_USAGE, base_fee)
 }
 
 pub fn aggregate_network_fee(
-    aggregate_size: i64,
+    aggregate_size: usize,
     gas_usage: &BigInt,
     base_fee: &TokenAmount,
 ) -> TokenAmount {
